Add test against OVS
The patch adds simple test where LLC information is passed to a Neutron
logical port. This triggers Ansible mechanism driver that is configured
to talk to local Open vSwitch database. The port on OVS side should be
configured with VLAN tag that is taken from Neutron logical network
segmentation ID. VLAN tag is validated using ovsdbapp querying Port OVS
table.

update_port method for openvswitch provider is added in order to be able
to talk to OVS DB.

NOTE: Current Ansible contains bug in openvswitch_db module, the PR is
      pending review at the time of pushing this patch at
      https://github.com/ansible/ansible/pull/42110

This is not a real tempest test as it assumes all-in-one and checks OVS
installed on the local machine from which tempest is run. This is
because we're in time pressure and such test adds a big value and there
is no better testing framework at this moment.

Change-Id: Ic73065175746a10ebef1ec731f947ee1c0919e72