tempest: Update list of files that should not trigger test
The patch removes requirements from the whitelist of files that don't
trigger tempest job in case there is a change in them. It's good to
trigger this job in case we bump some requirements in order to prevent
regressions.

The patch also updates path to unittests directory as it makes no sense
to trigger this tempest job when we make a change in unittests only.

Change-Id: Id1a957cd9c07a27b07c638f3890c38a0735b5bf8