Move unittests to unit directory
The patch introduces and official guideline for this project to keep
consistent naming of modules containing unittests. The structure is now
checked by the pep8 job.

The patch is not supposed to make any changes to class or test method
naming. It just squashes files together and meets the criteria for
unittests modules naming convention.

The only exception is renaming class
TestMechDriverPortChecks to TestMechDriverBindPort because of name
collision.

Change-Id: I0499dad8e9f14864636d0e8af5cd73d192b9ec01