Ensure all switch_mac refs are upper()
MAC map was not being used properly because the condition was not
forcing the mac to upper as expected. This moves the upper() func
to the call to get the initial value instead of running it later
when needed.

(cherry picked from commit d205a72267dc5b0147409f2771de6d012ef066a3)

Change-Id: Ic546bba5bf82842a6a4eb31fccc8f1b63a036fa3