Add a mac parameter to populate mac_map for introspection
Ironic introspection only provides the switch MAC and node's port
but ansible uses a name identifier to indicate which host to operate on.

This patch adds a mapping functionality that will populate the
ansible host value in playbooks when a switch's MAC is defined
in the ML2 ini for the device and the value matches the value provided
by lldp.

Change-Id: I489e06ce8a443dd0a96f0ee580847bb314640807