junos: Retry call to the switch if DB was locked
In case of concurrent requests to Neutron server ansible is not fast
enough to update ports on switch side and it happens two requests are
trying to update switch port at a time. Junos is not able to handle that
as it requires to commit changes before making any new changes.

This patch implements optimistic approach by retrying failed tasks in
case message about concurrency appears. This is specific to Junos and
hence goes to the provider directory.

Change-Id: I19b170d7b0ef5ec7f7c6690f6a07f7f9cda21639
Story: 2004409
Task: 28051