Remove LLI check if not needed
We don't need local link information for ports that are already bound.
In case subnet is created, Neutron automatically creates a DHCP port
when DHCP is enabled. Later, this port is passed to
update_port_postcommit() mech driver method for processing but doesn't
have any local link information.

This patch removes the check and keeps the check only for unbound ports,
thus ports that are already bound won't get this check.

Story: 2004793
Task: 28940

Change-Id: I2bca307a93e8301d2fc09be6ac40404215f0f7da